+++ Source/DotNetWorkQueue/Queue/ConsumerMetricsNotificationNoOp.cs
@@ -1,0 +1,10 @@
+namespace DotNetWorkQueue.Queue
+{
+    internal class ConsumerMetricsNotificationNoOp : IConsumerMetricsNotification
+    {
+        public void IncrementProcessed() { }
+        public void IncrementErrored() { }
+        public void IncrementRolledBack() { }
+        public void IncrementPoisonMessage() { }
+    }
+}