+++ Source/DotNetWorkQueue/Queue/ConsumerQueueErrorNotification.cs
@@ -4,9 +4,17 @@ namespace DotNetWorkQueue.Queue
 {
     internal class ConsumerQueueErrorNotification : IConsumerQueueErrorNotification
     {
+        private readonly IConsumerMetricsNotification _metrics;
         private ConsumerQueueNotifications _notifications;
+
+        public ConsumerQueueErrorNotification(IConsumerMetricsNotification metrics)
+        {
+            _metrics = metrics;
+        }
+
         public void InvokeError(ErrorNotification error)
         {
+            _metrics.IncrementErrored();
             _notifications?.Error?.Invoke(error);
         }
 
@@ -22,6 +30,7 @@ namespace DotNetWorkQueue.Queue
 
         public void InvokePoisonMessageError(PoisonMessageNotification notification)
         {
+            _metrics.IncrementPoisonMessage();
             _notifications?.PoisonMessage?.Invoke(notification);
         }